feat(docker): add support for reverse SSH tunnels in initramfs build
Introduce new build arguments for passing SSH keys and create a script to manage reverse tunnels. This enhances the Docker image's capabilities for remote access.

i want to be able to reboot this computer remotely but the hard disk with the persistent data is encrypted. how can i configure an initramfs that brings up the network, and starts an ssh server so that i can enter the password remotely. note that the system is behind cgnat. i have two servers outside (tinyeu, tinyus) and want to establish tunnels

follow the instructions below to implement this as a third initramfs in addition to the two that are already built by docker. the other images shall still work even if the new one has a problem:

Configuring an initramfs to unlock an encrypted root filesystem remotely is a robust way to manage servers that require both security and remote accessibility. Given your setup behind Carrier-Grade NAT (CGNAT) and the use of Gentoo with dracut, here is a detailed guide on how to achieve this.

The core of this solution involves using the dracut-crypt-ssh module, which integrates a lightweight SSH server (Dropbear) into the initramfs. This allows you to SSH into the pre-boot environment to provide the LUKS passphrase.